requirements.txt
    colorama==0.4.4
    Flask==2.0.1
    Flask-SocketIO==5.1.1
    Flask-Cors==3.0.10
    gunicorn==20.1.0
    eventlet==0.30.2
    requests==2.26.0
    python-dotenv==0.19.2